requirements.txt
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
ultralytics
opencv-python
streamlit
Pillow
matplotlib
numpy==1.26.4